+++ CHANGELOG.md
@@ -42,6 +42,8 @@ All notable changes to Rainbow-Node-SDK will be documented in this file.
 -   Update `InvitationsService::sendInvitationsByBulk` API with lang and comment parameters. 
 -   Add `InvitationsService::deleteAUserInvitation` API can be used to delete an invitation sent to/received from another Rainbow user. 
 -   Update `InvitationsService::reSendInvitation` API with customMessage parameters. 
+-   Add deleted and modified property in messages retrieved with history.
+-   Fix attention when user's jid_im is in mention array of an history message.
 
 ## [2.20.0] - 2023-02-24
 -   Add `PresenceService::setApplyMsTeamsPresenceSettings` This api allows to activate the exchange of presence of the connected user between rainbow and MS Teams on UI side.